Added gaussianFitter class and amended Gaussian class
The Gaussian class no longer relies on ParameterManager. Now the
parameters are managed by a specialised class called GaussianFitter,
which manages the names and values of the Gaussian. When the Gaussian is
copied parameter names are passed on.

Note: If two parameters are renamed to a single name then they can't be
disentangled.