Systematics can be associated with a group of EDs.
The SystematicManager now stores its systematics in map of vectors.
Each vector is referred to as a group and is keyed by the name of that
group. Systematics are added to a group through the BinnedNLLH (lh)
like:

    lh.AddSystematic(Systematic, "groupName");

The order in which the systematic is added to a group is the order in
which it is applied. If a group name is not specified, the systematic is
added the global ("") group which is then *always* applied to all EDs
before any other groups.

A ED is then associated with a vector of groups that will be applied to
that ED in that order:

    std::vector<std::string> appliedSys;
    appliedSys.push_back("GroupA");
    appliedSys.push_back("GroupB");
    lh.AddDist(ED,"appliedSys")

Global and group bases systematics may be applied together.

See examples/systematicFit.cpp for example fit.

Other changes

\#\#\# Details

There have been some changes to the to the previous PR.

\#\#\#\#\# Aim \#\#\#\#\#
We want to be able to add systematics to specific groups e.g.
"alphaEnergyRes", which can then be chained to each other and then
associated with a distribution.  However we also want to retain the
previous behaviour of adding global systematics to the lh which then
acts on all distributions. The problem with the previous attempt was
that it could, in some situations, become hard to know which groups were
applied.

To get around this I decided to change things so that in the following
ways:

* The global ("") group is now always created on initialisation and
given that it contains systematics is always applied to all distributions
in the lh.

* This change comes with some required extra logic:
  * When a distribution is added to the lh, the systematics it is
associated with are checked so that they don't apply to same systematic
twice and also checks the "" groups for the same reason.
  * When adding to the "" group the systematic is checked against all
existing groups to make sure that can't be applied twice.
  * In this way it is not possible for systematics to be applied twice.
  * There is then logic to apply the "" group before any other groups
associated with that distribution if it is populated.

* To get the correct ordering when calculating the total response matrix
for a group I changed the operator used. *Problem for anyone using
systematics in their fits*.

* The new scheme doesn't allow for a "group chain" applied to a
distribution to contain the "" group. e.g.
{"alphaEnergyRes","","somethingElse"}. Something like this can easily be
achieved by specifying the groups explicitly.
  * Also the person has to know about the "" group, which can only be
known by looking the code or this commit message. And therefore they are
being too clever for me and deserve everything they get. But honestly I
tried to put that in and it is so hard in this scheme.